tests/wpt/tests/html/canvas/element/drawing-images-to-the-canvas/2d.drawImage.svg.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.drawImage.svg</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.drawImage.svg</h1>
<p class="desc">drawImage() of an SVG image</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
promise_test(async t => {

  var canvas = document.getElementById('c');
  var ctx = canvas.getContext('2d');

  const img = new Image();
  const imageLoadPromise = new Promise((resolve, reject) => {
    img.onload = () => resolve();
    img.onerror = (err) => reject(err);
  });
  img.src = '/images/green.svg';
  await imageLoadPromise;

  ctx.drawImage(img, 0, 0);
  _assertPixelApprox(canvas, 50, 25, 0, 255, 0, 255, 2);

}, "drawImage() of an SVG image");
</script>
